authority_selection_inherents

Module filter_invalid_candidates

Functionality related to filtering invalid candidates from the candidates

Structs

CandidateWithStake

PermissionedCandidate

RegisterValidatorSignedMessage

Signed Message of the Authority Candidate to register It’s ToDatum implementation has to produce datum that has to match main chain structure, because it will be hashed and signed for signature verification.

Enums

Candidate

PermissionedCandidateDataError

RegistrationDataError

StakeError

Traits

CandidateValidationApi

Functions

filter_invalid_permissioned_candidates

filter_trustless_candidates_registrations

Get the valid trustless candidates from the registrations from inherent data

validate_permissioned_candidate_data

validate_registration_data

Is the registration data provided by the authority candidate valid?

validate_stake