authority_selection_inherents

Function select_authorities

pub fn select_authorities<TAccountId: Clone + Ord + From<Public>, TAccountKeys: Clone + Ord + MaybeFromCandidateKeys, MaxAuthorities: Get<u32>>(
    genesis_utxo: UtxoId,
    input: AuthoritySelectionInputs,
    sidechain_epoch: ScEpochNumber,
) -> Option<BoundedVec<CommitteeMember<TAccountId, TAccountKeys>, MaxAuthorities>>

Selects authorities using the Ariadne selection algorithm and data sourced from Partner Chains smart contracts on Cardano. Seed is constructed from the MC epoch nonce and the sidechain epoch.