Function authority_selection_inherents::select_authorities::select_authorities

pub fn select_authorities<TAccountId: Clone + Ord + TryFrom<SidechainPublicKey> + From<Public>, TAccountKeys: Clone + From<(Public, Public)>, Params: Clone + ToDatum, MaxValidators: Get<u32>>(
    sidechain_params: Params,
    input: AuthoritySelectionInputs,
    sidechain_epoch: ScEpochNumber,
) -> Option<BoundedVec<(TAccountId, TAccountKeys), MaxValidators>>

Pseudo-random selection the authorities for the given sidechain epoch, according to the Ariadne specification: https://input-output.atlassian.net/wiki/spaces/SID/pages/4228612151/Ariadne+-+committee+selection+algorithm

Seed is constructed from the MC epoch nonce and the sidechain epoch.

Committee size is P+T, where P (permissioned) and T (trustless) are constituents of the D parameter.

Committee is a result of the weighted selection with repetition.

Weight function for trustless candidate is:

let n be the number of permissioned candidates from MC data

if n == 0, then the weight is stake_delegation

otherwise, the weight is n * T * stake_delegation

Weight for each permissioned candidates is:

let W be the sum of all stake delegations of trustless candidates

if W == 0 or T == 0 (there are no valid trustless candidates, or they are not taken into account), then the weight is 1

otherwise, the weight is P * W