Crate authority_selection_inherents

This crate provides inherents for authority selection.

Modules

ariadne_inherent_data_provider

Logic to provide committee selection inputs compatible with the Ariadne algorithm, based on Cardano observation.

authority_selection_inputs

Types for authority selection

filter_invalid_candidates

Functionality related to filtering invalid candidates from the candidates

select_authorities

Functionality related to selecting the validators from the valid candidates

Enums

CommitteeMember

Type representing committee members, either permissioned or registered