Function authority_selection_inherents::select_authorities::select_authorities

pub fn select_authorities<TAccountId: Clone + Ord + TryFrom<SidechainPublicKey> + From<Public>, TAccountKeys: Clone + Ord + From<(Public, Public)>>(
    genesis_utxo: UtxoId,
    input: AuthoritySelectionInputs,
    sidechain_epoch: ScEpochNumber,
) -> Option<Vec<Candidate<TAccountId, TAccountKeys>>>

Selects authorities using the Ariadne selection algorithm and data sourced from Partner Chains smart contracts on Cardano. Seed is constructed from the MC epoch nonce and the sidechain epoch.